FitLens-dev2/requirements_rcnn.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
# Note: detectron2 must be installed separately (see installation instructions)

# Computer Vision
opencv-python==4.8.0
Pillow>=10.0.0
mediapipe==0.10.14
protobuf==4.25.3
onnxruntime==1.20.1
onnx==1.20.1
insightface

# Scientific Computing
numpy==1.23.5
scipy>=1.11.0

# Data Processing
pycocotools>=2.0.6
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Additional dependencies for detectron2
fvcore>=0.1.5
iopath>=0.1.7
omegaconf>=2.1.0
hydra-core>=1.1

# UI Framework
tkinter-tooltip>=2.1.0
customtkinter>=5.2.0

# Text-to-Speech
pyttsx3>=2.90

# Image Processing
scikit-image>=0.21.0
